requirements.txt
ipykernel
python-dotenv

assemblyai
deep-translator